web: files: transfer InitFilesStore to FilesStore
