+++ products/ASC.Files/Client/src/components/Article/Body/index.js
@@ -128,9 +128,8 @@ class ArticleBodyContent extends React.Component {
 }
 
 export default inject(
-  ({ initFilesStore, filesStore, treeFoldersStore, selectedFolderStore }) => {
-    const { setIsLoading } = initFilesStore;
-    const { fetchFiles, filter } = filesStore;
+  ({ filesStore, treeFoldersStore, selectedFolderStore }) => {
+    const { fetchFiles, filter, setIsLoading } = filesStore;
     const { treeFolders, setSelectedNode, setTreeFolders } = treeFoldersStore;
     const selectedTreeNode =
       treeFoldersStore.selectedTreeNode.length > 0
