+++ products/ASC.Files/Client/src/components/DragTooltip/index.js
@@ -4,6 +4,6 @@ import DragTooltip from "./DragTooltip";
 
 const Tooltip = ({ dragging }) => (dragging ? <DragTooltip /> : <></>);
 
-export default inject(({ initFilesStore, filesStore }) => ({
-  dragging: initFilesStore.dragging && filesStore.selection[0],
+export default inject(({ filesStore }) => ({
+  dragging: filesStore.dragging && filesStore.selection[0],
 }))(observer(Tooltip));
