+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/EmptyContainer/EmptyFolderContainer.js
@@ -79,13 +79,13 @@ const EmptyFolderContainer = ({
   );
 };
 
-export default inject(({ initFilesStore, filesStore, selectedFolderStore }) => {
+export default inject(({ filesStore, selectedFolderStore }) => {
   const { filter, fetchFiles } = filesStore;
 
   return {
     filter,
     fetchFiles,
-    setIsLoading: initFilesStore.setIsLoading,
+    setIsLoading: filesStore.setIsLoading,
     parentId: selectedFolderStore.parentId,
   };
 })(withTranslation("Home")(observer(EmptyFolderContainer)));
