+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/EmptyContainer/RootFolderContainer.js
@@ -188,21 +188,19 @@ const RootFolderContainer = (props) => {
 };
 
 export default inject(
-  ({
-    auth,
-    initFilesStore,
-    filesStore,
-    treeFoldersStore,
-    selectedFolderStore,
-  }) => {
+  ({ auth, filesStore, treeFoldersStore, selectedFolderStore }) => {
     const {
       isDesktopClient,
       isEncryptionSupport,
       organizationName,
     } = auth.settingsStore;
 
-    const { privacyInstructions, setIsLoading } = initFilesStore;
-    const { filter, fetchFiles } = filesStore;
+    const {
+      filter,
+      fetchFiles,
+      privacyInstructions,
+      setIsLoading,
+    } = filesStore;
     const { title, rootFolderType } = selectedFolderStore;
     const { isPrivacyFolder, myFolderId } = treeFoldersStore;
