--- products/ASC.Files/Client/src/store/InitFilesStore.js
@@ -1,120 +1,0 @@
-import { makeAutoObservable } from "mobx";
-import { updateTempContent } from "@appserver/common/utils";
-import config from "../../package.json";
-
-class InitFilesStore {
-  authStore;
-  settingsStore;
-  filesStore;
-  treeFoldersStore;
-
-  isLoaded = false;
-  isLoading = false;
-  viewAs = "row";
-  dragging = false;
-  privacyInstructions = "https://www.onlyoffice.com/private-rooms.aspx";
-  isInit = false;
-
-  tooltipPageX = 0;
-  tooltipPageY = 0;
-
-  constructor(authStore, settingsStore, filesStore, treeFoldersStore) {
-    const pathname = window.location.pathname.toLowerCase();
-    this.isEditor = pathname.indexOf("doceditor") !== -1;
-
-    makeAutoObservable(this);
-    this.authStore = authStore;
-    this.settingsStore = settingsStore;
-    this.filesStore = filesStore;
-    this.treeFoldersStore = treeFoldersStore;
-  }
-
-  setIsLoaded = (isLoaded) => {
-    this.isLoaded = isLoaded;
-  };
-
-  setViewAs = (viewAs) => {
-    this.viewAs = viewAs;
-  };
-
-  setDragging = (dragging) => {
-    this.dragging = dragging;
-  };
-
-  setIsLoading = (isLoading) => {
-    this.isLoading = isLoading;
-  };
-
-  setTooltipPosition = (tooltipPageX, tooltipPageY) => {
-    this.tooltipPageX = tooltipPageX;
-    this.tooltipPageY = tooltipPageY;
-  };
-
-  get tooltipValue() {
-    if (!this.dragging) return null;
-
-    const selectionLength = this.filesStore.selection.length;
-    const elementTitle = selectionLength && this.filesStore.selection[0].title;
-    const singleElement = selectionLength === 1;
-    const filesCount = singleElement ? elementTitle : selectionLength;
-    const { isShareFolder, isCommonFolder } = this.treeFoldersStore;
-
-    let operationName;
-
-    if (this.authStore.isAdmin && isShareFolder) {
-      operationName = "copy";
-    } else if (!this.authStore.isAdmin && (isShareFolder || isCommonFolder)) {
-      operationName = "copy";
-    } else {
-      operationName = "move";
-    }
-
-    return operationName === "copy"
-      ? singleElement
-        ? { label: "TooltipElementCopyMessage", filesCount }
-        : { label: "TooltipElementsCopyMessage", filesCount }
-      : singleElement
-      ? { label: "TooltipElementMoveMessage", filesCount }
-      : { label: "TooltipElementsMoveMessage", filesCount };
-  }
-
-  initFiles = () => {
-    if (this.isInit) return;
-    this.isInit = true;
-
-    const { isAuthenticated } = this.authStore;
-    const {
-      getPortalCultures,
-      isDesktopClient,
-      getIsEncryptionSupport,
-      getEncryptionKeys,
-      setModuleInfo,
-    } = this.settingsStore;
-
-    setModuleInfo(config.homepage, config.id);
-
-    const requests = [];
-
-    updateTempContent();
-    if (!isAuthenticated) {
-      return this.setIsLoaded(true);
-    } else {
-      updateTempContent(isAuthenticated);
-    }
-
-    if (!this.isEditor) {
-      requests.push(
-        getPortalCultures(),
-        this.treeFoldersStore.fetchTreeFolders()
-      );
-
-      if (isDesktopClient) {
-        requests.push(getIsEncryptionSupport(), getEncryptionKeys());
-      }
-    }
-
-    return Promise.all(requests);
-  };
-}
-
-export default InitFilesStore;
